add back plmsheetlistmodel.cpp on core ; fix qml access in main.cpp; delete plmuserfilehub.cpp cocs of merge in mai db; fill plmuserhub.cpp with basics

+++ src/app/src/documenthandler.h
@@ -5,16 +5,20 @@
 #include <QQuickTextDocument>
 #include <QTextCursor>
 
-class DocumentHandler: public QObject
-{
+class DocumentHandler : public QObject {
     Q_OBJECT
     Q_PROPERTY(QStringList allFontFamilies READ allFontFamilies CONSTANT)
 
-    Q_PROPERTY(QQuickTextDocument* textDocument READ textDocument WRITE setTextDocument NOTIFY textDocumentChanged)
+    Q_PROPERTY(
+        QQuickTextDocument *
+        textDocument READ textDocument WRITE setTextDocument NOTIFY textDocumentChanged)
 
-    Q_PROPERTY(int cursorPosition READ cursorPosition WRITE setCursorPosition NOTIFY cursorPositionChanged)
-    Q_PROPERTY(int selectionStart READ selectionStart WRITE setSelectionStart NOTIFY cursorPositionChanged)
-    Q_PROPERTY(int selectionEnd READ selectionEnd WRITE setSelectionEnd NOTIFY cursorPositionChanged)
+    Q_PROPERTY(
+        int cursorPosition READ cursorPosition WRITE setCursorPosition NOTIFY cursorPositionChanged)
+    Q_PROPERTY(
+        int selectionStart READ selectionStart WRITE setSelectionStart NOTIFY cursorPositionChanged)
+    Q_PROPERTY(
+        int selectionEnd READ selectionEnd WRITE setSelectionEnd NOTIFY cursorPositionChanged)
 
     Q_PROPERTY(QString fontFamily READ fontFamily WRITE setFontFamily NOTIFY formatChanged)
     Q_PROPERTY(qreal fontSize READ fontSize WRITE setFontSize NOTIFY formatChanged)
@@ -23,9 +27,11 @@ class DocumentHandler: public QObject
     Q_PROPERTY(bool underline READ underline WRITE setUnderline NOTIFY formatChanged)
     Q_PROPERTY(bool strikeout READ strikeout WRITE setStrikeout NOTIFY formatChanged)
     Q_PROPERTY(QColor color READ color WRITE setColor NOTIFY formatChanged)
-    Q_PROPERTY(Qt::Alignment alignment READ alignment WRITE setAlignment NOTIFY formatChanged)
+    Q_PROPERTY(
+        Qt::Alignment alignment READ alignment WRITE setAlignment NOTIFY formatChanged)
     Q_PROPERTY(bool bulletList READ bulletList WRITE setBulletList NOTIFY formatChanged)
-    Q_PROPERTY(bool numberedList READ numberedList WRITE setNumberedList NOTIFY formatChanged)
+    Q_PROPERTY(
+        bool numberedList READ numberedList WRITE setNumberedList NOTIFY formatChanged)
 
     Q_PROPERTY(bool canUndo READ canUndo NOTIFY canUndoChanged)
     Q_PROPERTY(bool canRedo READ canRedo NOTIFY canRedoChanged)
@@ -33,62 +39,64 @@ class DocumentHandler: public QObject
     Q_PROPERTY(int projectId READ projectId NOTIFY idChanged)
 
 public:
+
     DocumentHandler(QObject *parent = nullptr);
 
-    QStringList allFontFamilies() const;
+    QStringList         allFontFamilies() const;
 
     QQuickTextDocument* textDocument() const;
-    void setTextDocument(QQuickTextDocument* textDocument);
+    void                setTextDocument(QQuickTextDocument *textDocument);
 
-    int cursorPosition() const;
-    void setCursorPosition(int position);
+    int                 cursorPosition() const;
+    void                setCursorPosition(int position);
 
-    int selectionStart() const;
-    void setSelectionStart(int selectionStart);
+    int                 selectionStart() const;
+    void                setSelectionStart(int selectionStart);
 
-    int selectionEnd() const;
-    void setSelectionEnd(int selectionEnd);
+    int                 selectionEnd() const;
+    void                setSelectionEnd(int selectionEnd);
 
-    QString fontFamily() const;
-    void setFontFamily(const QString &fontFamily);
+    QString             fontFamily() const;
+    void                setFontFamily(const QString& fontFamily);
 
-    int fontSize() const;
-    void setFontSize(qreal fontSize);
+    qreal               fontSize() const;
+    void                setFontSize(qreal fontSize);
 
-    bool italic() const;
-    void setItalic(bool italic);
+    bool                italic() const;
+    void                setItalic(bool italic);
 
-    bool bold() const;
-    void setBold(bool bold);
+    bool                bold() const;
+    void                setBold(bool bold);
 
-    bool underline() const;
-    void setUnderline(bool underline);
+    bool                underline() const;
+    void                setUnderline(bool underline);
 
-    bool strikeout() const;
-    void setStrikeout(bool strikeout);
+    bool                strikeout() const;
+    void                setStrikeout(bool strikeout);
 
-    QColor color() const;
-    void setColor(const QColor &color);
+    QColor              color() const;
+    void                setColor(const QColor& color);
 
-    bool canUndo() const;
-    bool canRedo() const;
+    bool                canUndo() const;
+    bool                canRedo() const;
 
-    Qt::Alignment alignment() const;
-    void setAlignment(Qt::Alignment alignment);
+    Qt::Alignment       alignment() const;
+    void                setAlignment(Qt::Alignment alignment);
 
-    bool bulletList() const;
-    void setBulletList(bool bulletList);
+    bool                bulletList() const;
+    void                setBulletList(bool bulletList);
 
-    bool numberedList() const;
-    void setNumberedList(bool numberedList);
+    bool                numberedList() const;
+    void                setNumberedList(bool numberedList);
 
-    void setId(const int projectId, const int paperId);
-    int paperId() const;
+    void                setId(const int projectId,
+                              const int paperId);
+    int                 paperId() const;
 
-    int projectId() const;
-
+    int                 projectId() const;
 
 public slots:
+
     void addHorizontalLine();
     void indentBlock();
     void unindentBlock();
@@ -97,6 +105,7 @@ public slots:
     void redo();
 
 signals:
+
     void textDocumentChanged();
     void cursorPositionChanged();
     void formatChanged();
@@ -106,6 +115,7 @@ signals:
     void idChanged();
 
 private:
+
     QQuickTextDocument *m_textDoc;
     QTextCursor m_textCursor;
     QTextCursor m_selectionCursor;
@@ -116,7 +126,6 @@ private:
     int m_selectionEnd;
     int m_projectId;
     int m_paperId;
-
 };
 
 #endif // DOCUMENTHANDLER_H
